04 auth

+++ controllers/contactsControllers.js
@@ -3,8 +3,18 @@ import HttpError from "../helpers/HttpError.js";
 
 export const getAllContacts = async (req, res, next) => {
   try {
-    const data = await contactsService.listContacts();
-    res.json(data);
+    const { page = 1, limit = 10, favorite } = req.query;
+    const { _id: owner } = req.user;
+    const skip = (page - 1) * limit;
+    const filter = favorite !== undefined ? { owner, favorite } : { owner };
+    // const filter = Object.assign(
+    //   { owner },
+    //   favorite !== undefined && { favorite }
+    // );
+    const data = await contactsService.listContacts(filter, { skip, limit });
+    const total = await contactsService.countContacts({ owner });
+    const perpage = data.length;
+    res.json({ data, total, perpage });
   } catch (error) {
     next(error);
   }
@@ -12,8 +22,9 @@ export const getAllContacts = async (req, res, next) => {
 
 export const getOneContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { id } = req.params;
-    const data = await contactsService.getContactById(id);
+    const data = await contactsService.getContactById({ owner, id });
     if (!data) throw HttpError(404);
     res.json(data);
   } catch (error) {
@@ -24,7 +35,8 @@ export const getOneContact = async (req, res, next) => {
 export const deleteContact = async (req, res, next) => {
   try {
     const { id } = req.params;
-    const data = await contactsService.removeContact(id);
+    const { _id: owner } = req.user;
+    const data = await contactsService.removeContact({ owner, id });
     if (!data) throw HttpError(404);
     res.json(data);
   } catch (error) {
@@ -34,7 +46,8 @@ export const deleteContact = async (req, res, next) => {
 
 export const createContact = async (req, res, next) => {
   try {
-    const data = await contactsService.addContact(req.body);
+    const { _id: owner } = req.user;
+    const data = await contactsService.addContact({ ...req.body, owner });
     res.status(201).json(data);
   } catch (error) {
     next(error);
@@ -46,7 +59,11 @@ export const updateContact = async (req, res, next) => {
     const isBodyEmpty = Object.keys(req.body).length === 0;
     if (isBodyEmpty) throw HttpError(400, "Body must have at least one field");
     const { id } = req.params;
-    const data = await contactsService.updateContact(id, req.body);
+    const { _id: owner } = req.user;
+    const data = await contactsService.updateContact(
+      { owner, _id: id },
+      req.body
+    );
     if (!data) throw HttpError(404);
     res.json(data);
   } catch (error) {
@@ -55,10 +72,11 @@ export const updateContact = async (req, res, next) => {
 };
 export const updateStatusContact = async (req, res, next) => {
   try {
+    const { _id: owner } = req.user;
     const { id } = req.params;
     const { favorite } = req.body;
     const updateFavorite = await contactsService.updateStatusContact(
-      id,
+      { owner, _id: id },
       favorite
     );
     if (!updateFavorite) throw HttpError(404, "message not found");
